Remove v2 asset upload strategies and fix tests for updated data-model API
